+++ src/disk/sys/macos/disk_io_counters.rs
@@ -1,8 +1,8 @@
 use std::collections::HashMap;
-use std::io;
 
 use crate::disk::DiskIoCounters;
+use crate::Result;
 
-pub(crate) fn disk_io_counters_per_partition() -> io::Result<HashMap<String, DiskIoCounters>> {
+pub(crate) fn disk_io_counters_per_partition() -> Result<HashMap<String, DiskIoCounters>> {
 	todo!()
 }