+++ src/memory/sys/macos/mod.rs
@@ -2,5 +2,6 @@ mod common;
 mod swap_memory;
 mod virtual_memory;
 
+pub(crate) use common::*;
 pub use swap_memory::*;
 pub use virtual_memory::*;