+++ src/network/sys/linux/net_io_counters.rs
@@ -1,48 +1,61 @@
 use std::collections::HashMap;
-use std::fs;
-use std::io;
 use std::str::FromStr;
 
+use snafu::{ensure, ResultExt};
+
 use crate::network::NetIoCounters;
-use crate::utils::invalid_data;
+use crate::{read_file, Error, MissingData, ParseInt, Result};
+
+const PROC_NET_DEV: &str = "/proc/net/dev";
 
 impl FromStr for NetIoCounters {
-	type Err = std::io::Error;
+	type Err = Error;
 
-	fn from_str(line: &str) -> Result<Self, Self::Err> {
+	fn from_str(line: &str) -> Result<Self> {
 		let fields: Vec<&str> = line.split_whitespace().collect();
 
-		if fields.len() < 17 {
-			return Err(invalid_data(
-				"'/proc/net/dev' does not have the right number of fields",
-			));
-		}
+		ensure!(
+			fields.len() >= 17,
+			MissingData {
+				path: PROC_NET_DEV,
+				contents: line,
+			}
+		);
+
+		let parse = |s: &str| -> Result<u64> {
+			s.parse().context(ParseInt {
+				path: PROC_NET_DEV,
+				contents: line,
+			})
+		};
 
 		Ok(NetIoCounters {
-			bytes_sent: try_parse!(fields[9]),
-			bytes_recv: try_parse!(fields[1]),
-			packets_sent: try_parse!(fields[10]),
-			packets_recv: try_parse!(fields[2]),
-			err_in: try_parse!(fields[3]),
-			err_out: try_parse!(fields[11]),
-			drop_in: try_parse!(fields[4]),
-			drop_out: try_parse!(fields[12]),
+			bytes_sent: parse(fields[9])?,
+			bytes_recv: parse(fields[1])?,
+			packets_sent: parse(fields[10])?,
+			packets_recv: parse(fields[2])?,
+			err_in: parse(fields[3])?,
+			err_out: parse(fields[11])?,
+			drop_in: parse(fields[4])?,
+			drop_out: parse(fields[12])?,
 		})
 	}
 }
 
-pub(crate) fn net_io_counters_pernic() -> io::Result<HashMap<String, NetIoCounters>> {
-	fs::read_to_string("/proc/net/dev")?
+pub(crate) fn net_io_counters_pernic() -> Result<HashMap<String, NetIoCounters>> {
+	read_file(PROC_NET_DEV)?
 		.lines()
 		.skip(2)
 		.map(|line| {
 			let fields: Vec<&str> = line.split_whitespace().collect();
 
-			if fields.len() < 17 {
-				return Err(invalid_data(
-					"'/proc/net/dev' does not have the right number of fields",
-				));
-			}
+			ensure!(
+				fields.len() >= 17,
+				MissingData {
+					path: PROC_NET_DEV,
+					contents: line,
+				}
+			);
 
 			let mut net_name = String::from(fields[0]);
 			// remove the trailing colon