+++ src/process/sys/linux/pids.rs
@@ -1,15 +1,13 @@
-use std::fs;
-use std::io;
 use std::path::Path;
 
-use crate::Pid;
+use crate::{read_dir, Pid, Result};
 
 // TODO: should we return an `io::Result<Vec<io::Result<<Pid>>>` instead?
-pub fn pids() -> io::Result<Vec<Pid>> {
+pub fn pids() -> Result<Vec<Pid>> {
 	let mut pids = Vec::new();
 
-	for entry in fs::read_dir("/proc")? {
-		let filename = entry?.file_name();
+	for entry in read_dir("/proc")? {
+		let filename = entry.file_name();
 		if let Ok(pid) = filename.to_string_lossy().parse::<Pid>() {
 			pids.push(pid);
 		}